Skip field removed, icons changed, marquee changed
I've removed the text entry field from the skip buttons, they now skip based on a skipValue variable, currently 5 seconds.

Changed the skip icons to be a tiny bit less confusing, but there might be better icons out there.

Sorted the marquee out a little bit, still not entirely satisfied with it.

I tried to look into the playing next issue, but I can't find any relevant documentation for the elements of the Youtube player, I'm not sure where the names for elements being retrieved for in PageListener.js are coming from.