Current Tab Updates
Current Tab now persists through opening/closing the addon

Current Tab can be moved between tabs in position (e.g the current tab is the 2nd out of 4 tabs, if it is moved to be the 1st out of 4 it will still play the moved video)

Current tab shows in dropdown accurately. This currently works by firing off a timer about 250ms after the addon is opened. I'm not too happy about how this works, I'd like this to work a little bit neater, like work based off when the dropdown is ready, as just calling for the selected item in the dropdown before its ready (like at the end of main) casues it to just not work .

Also removed some unnecessary console.logs that were getting in the way