Current Tab Additions
Current Tab now persists and is shown when the video addon is loaded. The tabs can be shuffled around, the addon will pick the last selected tab, and display it in the drop-down, and the selected tab can be closed which will cause it to defaul to the first available tab.

Fixed some minor issues with information not being loaded when the addon was loaded.

Again, this all works fine on my computer. Feel free to undo it if it busts everything up like last time (and the time before).